Timer works as intended now
The timer works as intended now; I accomplished this using a global scene that tracks the number of bateyes killed which is kind of bad coding wise and if i can find a better way I'll do that because the player stats are also currently global. Not good practice for future. But the timer does stop when you die or when you kill every enemy so for now this is fine. Next I need a counter on how many bateyes are dead which is pretty easy at this point.